[refactor] Merge taichi/lang/linalg_impl.py into _funcs.py
[ghstack-poisoned]
